+++ lib/halbuilder/helper.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require_relative "./zoom"
+
+module Halbuilder::Helper
+  include Halbuilder::Zoom
+
+  def zoom_param
+    params[:zoom]
+  end
+end