+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-## [Unreleased]
+## [0.1.1] - 2022-07-21
+
+- Add `Halbuilder::Helper` module, automatically included in Rails views/controllers.
+  Currently only has a `hal_zoomed?` method to help decide when to preload associations.
+- `Jbuilder.ignore_nil(true)` by default
 
 ## [0.1.0] - 2022-07-20